+++ functions/public/Convert-ScriptToFunction.ps1
@@ -1,45 +1,4 @@
 Function Convert-ScriptToFunction {
-    <#
-    .Synopsis
-    Convert a script file to a PowerShell function.
-    .Description
-    This command takes the body of a script file and wraps it in a function
-    declaration. The command will insert missing elements like [cmdletbinding()]
-    and comment-based help. You will most likely need to edit and clean up the
-    result in your scripting editor.
-
-    If you run this command in the PowerShell ISE or the VS Code PowerShell
-    integrated terminal, you can use the dynamic parameter ToEditor to open a
-    new file with with the output. You can edit and save the file manually.
-
-    It is assumed that your script file is complete and without syntax errors.
-    .Parameter Path
-    Enter the path to your PowerShell script file.
-    .Parameter Name
-    What is the name of your new function? It should have a Verb-Noun name.
-    .Parameter Alias
-    Define an optional alias for your new function.
-    .Parameter ToEditor
-    If you run this command in the PowerShell ISE or the VS Code PowerShell
-    integrated terminal, you can use this dynamic parameter to open a new
-    file with with the output. You can edit and save the file manually.
-    .Example
-    PS C:\> Convert-ScriptToFunction c:\scripts\Daily.ps1 -name Invoke-DailyTask | Set-Clipboard
-
-    Convert Daily.ps1 to a function called Invoke-DailyTask and copy the
-    results to the Windows clipboard. You can then paste the results into
-    scripting editor.
-    .Example
-    PS C:\> Convert-ScriptToFunction c:\scripts\systemreport.ps1 -name New-SystemReport | Out-File c:\scripts\New-SystemReport.ps1
-
-    Convert the SystemReport.ps1 script file to a function called
-    New-SystemReport and save the results to a file.
-    .Example
-    PS C:\> Convert-ScriptToFunction c:\scripts\systemreport.ps1 -name New-System -alias nsr | Tee-Object -variable f
-
-    Convert the script to a function called New-System and tee the output to $f.
-    This will also define an function alias of nsr.
-    #>
     [cmdletbinding()]
     [Outputtype("System.String")]
     [alias('csf')]