+++ functions/public/Get-FunctionAlias.ps1
@@ -1,10 +1,10 @@
 Function Get-FunctionAlias {
     [cmdletbinding()]
-    [alias("gfal","ga")]
+    [alias("gfal", "ga")]
     [outputType("string")]
     Param(
         [Parameter(Position = 0, Mandatory, HelpMessage = "Specify the .ps1 or .psm1 file with defined functions.")]
-        [ValidateScript({
+        [ValidateScript( {
             If (Test-Path $_ ) {
                 $True
             }
@@ -13,7 +13,7 @@ Function Get-FunctionAlias {
                 $False
             }
         })]
-        [ValidateScript({
+    [ValidateScript( {
             If ($_ -match "\.ps(m)?1$") {
                 $True
             }
@@ -28,18 +28,25 @@ Function Get-FunctionAlias {
     New-Variable astTokens -Force
     New-Variable astErr -Force
     $Path = Convert-Path -Path $path
-    [regex]$rx = "(?<=alias\().*(?=\)\])"
     Write-Verbose "Parsing $path for functions."
     $AST = [System.Management.Automation.Language.Parser]::ParseFile($Path, [ref]$astTokens, [ref]$astErr)
 
     #parse out functions using the AST
-    $functions = $ast.FindAll({ $args[0] -is [System.Management.Automation.Language.FunctionDefinitionAst] }, $true)
+    $functions = $ast.FindAll( { $args[0] -is [System.Management.Automation.Language.FunctionDefinitionAst] }, $true)
     if ($functions.count -gt 0) {
         foreach ($f in $functions) {
-            if ($rx.IsMatch($f.body)) {
+            #thanks to https://twitter.com/PrzemyslawKlys for this suggestion
+            $aliasAST = $f.FindAll( {
+                    $args[0] -is [System.Management.Automation.Language.AttributeAst] -and
+                    $args[0].TypeName.Name -eq 'Alias' -and
+                    $args[0].Parent -is [System.Management.Automation.Language.ParamBlockAst]
+                }, $true)
+
+            if ($aliasAST.positionalArguments) {
                 [pscustomobject]@{
-                    Name = $f.name
-                    Alias = ($rx.matches($f.body).value -replace """|'","") -split ","
+                    PSTypeName = "PSFunctionAlias"
+                    Name       = $f.name
+                    Alias      = $aliasAST.PositionalArguments.Value
                 }
             }
         }