+++ changelog.md
@@ -1,1 +1,8 @@
-# Changelog
+# Changelog for PSFunctionTools
+
+## v0.1.0
+
++ Initial files.
++ Added `Get-ModuleLayout` with format file `modulelayout.format.ps1xml`.
++ Modified `Get-PSScriptRequirements` to write a `PSScriptRequirements` object to the pipeline. Added the format file `psscriptrequirements.format.ps1xml`.
++ Updated help documentation.